fix: correct Doctrine config, price type and Docker dev setup
- Fix MigrationsBundle class name in bundles.php
- Change Product price from decimal to float to match Doctrine mapping
- Add composer install to Dockerfile dev stage
- Add vendor named volume to docker-compose to preserve dependencies
- Update migration to use DOUBLE PRECISION for price column

+++ services/mini-allegro/migrations/Version20260315000001.php
@@ -24,7 +24,7 @@ final class Version20260315000001 extends AbstractMigration
                 id INT NOT NULL,
                 name VARCHAR(255) NOT NULL,
                 description TEXT DEFAULT NULL,
-                price NUMERIC(10, 2) NOT NULL,
+                price DOUBLE PRECISION NOT NULL,
                 PRIMARY KEY(id)
             )
         ');
